modified CSS dependency stack, configs, upgraded storybook...
