+++ .storybook/preview.js
@@ -1,6 +1,4 @@
 import React, { Component, Fragment } from 'react';
-import { addDecorator, addParameters } from '@storybook/react';
-import { withOptions } from '@storybook/addon-options';
 import { addReadme } from 'storybook-readme';
 import { initializeRTL } from 'storybook-addon-rtl';
 import '../lib/global.css';
@@ -58,7 +56,7 @@ setIntlConfig({
     getMessages: (locale) => messages[locale]
 });
 
-addDecorator(withIntl);
+// addDecorator(withIntl);
 
 /**
  * RTL
@@ -90,26 +88,41 @@ initializeRTL();
   }
  }
 
-addDecorator(storyFn => <AddOverlayContainer>{storyFn()}</AddOverlayContainer>);
+// addDecorator(storyFn => <AddOverlayContainer>{storyFn()}</AddOverlayContainer>);
 
 /**
  * Set options
  */
-addDecorator(withOptions({
-  name: 'FOLIO Stripes',
-  // sortStoriesByKind:
-  hierarchySeparator: /\|/,
-  hierarchyRootSeparator: /\|/,
-}));
+// addDecorator(withOptions({
+//   name: 'FOLIO Stripes',
+//   // sortStoriesByKind:
+//   hierarchySeparator: /\|/,
+//   hierarchyRootSeparator: /\|/,
+// }));
 
-addParameters({ options: { theme: {} } });
+// addParameters({ options: { theme: {} } });
 
 /**
  * Readme
  */
-addDecorator(addReadme);
-addParameters({ 
-    readme: {
-        codeTheme: 'a11y-dark',
-    },
- });
+// addDecorator(addReadme);
+// addParameters({ 
+//     readme: {
+//         codeTheme: 'a11y-dark',
+//     },
+//  });
+
+
+const storyFnDecorator = storyFn => <AddOverlayContainer>{storyFn()}</AddOverlayContainer>
+
+ export const decorators = [withIntl, storyFnDecorator, addReadme]
+ export const parameters = {
+   options: {
+     theme: {
+      brandTitle: 'FOLIO Stripes',
+     }
+   },
+   readme: {
+     codeTheme: 'a11y-dark',
+   }
+ }
