+++ docs/AccessibilityDevPrimer.md
@@ -47,7 +47,7 @@ After a user has finished working within the modal, focus should travel back/clo
 Programmatic focus changes should only happen **after a user triggers them**. As a rule, focusing one element should not cause focus to move to another([3.2.1](https://www.w3.org/TR/WCAG21/#on-focus)).
 
 In ReactJS, focus can be managed via `ref`s to DOM elements. In this contrived example, an element is programmatically focused by calling the element's `focus()` method via the ref:
-```jsx static
+```
 this.addButton = react.createRef();
 
 focusAddButton = () => {
@@ -81,7 +81,7 @@ Color contrast is an important aspect to ensure that the webpage is legible, eve
 Form controls require associated labels so that AT can convey the purpose of a text field, select box, radio-button, etc.([1.3.5](https://www.w3.org/TR/WCAG21/#identify-input-purpose)) While our components do encapsulate numerous accessibility features, labeling is one aspect where options are preserve due to the variance in techniques.
 
 Typical labeling of form elements is as follows (explicit labeling) our form controls set these up under the hood:
-```jsx static
+```
 <!-- general html controls -->
 <label for="inputid">username</label>
 <input type="text" id="inputid" />
@@ -93,7 +93,7 @@ Typical labeling of form elements is as follows (explicit labeling) our form con
 Given this, a screen reader will announce 'username edit blank' when the user tabs to the field. Without the `<label>`, it simply announces 'edit blank'.
 
 If visible labels are inadequate, `aria-label` can be used:
-```jsx static
+```
 <!-- general html controls -->
 <input type="text" aria-label="username" id="inputid" />
 
@@ -101,7 +101,7 @@ If visible labels are inadequate, `aria-label` can be used:
 <TextField aria-label="username" />
 ```
 If the label is visible elsewhere in the page, its element can be associated with the input via `aria-labelledby`: 
-```jsx static
+```
 <!-- general html controls -->
 <div id="columnlabel">starting date</div>
 ...
@@ -123,7 +123,7 @@ It is not uncommon for `title` attributes to be used and considered 'handy' for 
 
 In the following contrived example, we might have wanted the label to be 'username' - but the `<label>` element will be ignored despite its explicit relationship, as well as the `aria-label` attribute.
 The content of the `<span #label-elem>` will be announced, as well as the `title`. Focusing this input using NVDA will announce 'label element edit title name, blank'.
-```jsx static
+```
 <span id="label-elem">label element</span>
 <TextField
   label="username"
@@ -164,7 +164,7 @@ The ideal document structure, at the very least, has headers nested by their lev
 ```
 Assistive technology allows users to navigate from one header to the next. Scanning headers is one of the first steps a screen reader user may take when they're getting acquainted with the markup of a page.
 Know that semantic landmark `<section>` can restart the heading rank order...
-```jsx static
+```
 <section>
   <H2> I'm an h2, but I'm considered an H1 due to the section tag </H2>
 </section>
