+++ package.json
@@ -48,12 +48,12 @@
     "@bigtest/mocha": "^0.5.0",
     "@folio/eslint-config-stripes": "^5.2.0",
     "@folio/stripes-cli": "^2.0.0",
-    "@storybook/addon-actions": "^5.3.19",
-    "@storybook/addon-knobs": "^5.3.19",
-    "@storybook/addon-options": "^5.3.19",
-    "@storybook/addons": "^5.3.19",
-    "@storybook/react": "^5.3.19",
-    "autoprefixer": "^7.1.6",
+    "@mdx-js/loader": "^1.6.22",
+    "@storybook/addon-actions": "^6.3.6",
+    "@storybook/addon-knobs": "^6.2.9",
+    "@storybook/addons": "^6.3.6",
+    "@storybook/react": "^6.3.6",
+    "autoprefixer": "^9.0.0",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.0.0",
     "babel-plugin-lodash": "^3.3.4",
@@ -68,14 +68,15 @@
     "karma-webpack": "^4.0.2",
     "mocha": "^5.2.0",
     "moment": "^2.29.0",
-    "postcss": "^7.0.2",
-    "postcss-calc": "^6.0.1",
-    "postcss-color-function": "^4.0.1",
-    "postcss-custom-media": "^6.0.0",
+    "postcss": "^7.0.0",
+    "postcss-calc": "^7.0.0",
+    "postcss-color-mod-function": "^3.0.0",
+    "postcss-custom-media": "^7.0.0",
     "postcss-custom-properties": "^10.0.0",
-    "postcss-import": "^11.0.0",
-    "postcss-media-minmax": "^3.0.0",
-    "postcss-nesting": "^4.2.1",
+    "postcss-import": "^12.0.1",
+    "postcss-loader": "^3.0.0",
+    "postcss-media-minmax": "^4.0.0",
+    "postcss-nesting": "^7.0.1",
     "postcss-syntax": "^0.34.0",
     "react": "^16.3.0",
     "react-dom": "^16.3.0",
@@ -87,10 +88,10 @@
     "redux-form": "^8.3.0",
     "regenerator-runtime": "^0.13.3",
     "sinon": "^7.3.2",
-    "storybook-addon-designs": "^5.4.2",
+    "storybook-addon-designs": "^6.0.1",
     "storybook-addon-intl": "^2.4.1",
-    "storybook-addon-rtl": "^0.2.2",
-    "storybook-readme": "^5.0.8",
+    "storybook-addon-rtl": "^0.3.0",
+    "storybook-readme": "^5.0.9",
     "stylelint": "^9.5.0",
     "stylelint-config-standard": "^18.2.0",
     "stylelint-junit-formatter": "^0.2.1",
